Move add item functionality into list view

+++ lists/views.py
@@ -22,18 +22,15 @@ def new_list(request):
 
 def view_list(request, list_id):
     list_ = List.objects.get(id=list_id)
-    items = Item.objects.filter(list=list_)
-    return render(request, 'list.html', {'list': list_})
+    error = None
 
-def add_item(request, list_id):
-    list_ = List.objects.get(id=list_id)
-
-    item = Item.objects.create(text=request.POST['item_text'], list=list_)
-    try:
-        item.full_clean()
-        item.save()
-    except ValidationError:
-        item.delete()
-        error = "You can't have an empty list item"
-        return render(request, 'list.html', {'list': list_, 'error': error})
-    return redirect(f'/lists/{list_.id}/')
+    if request.method == 'POST':
+        try:
+            item = Item.objects.create(text=request.POST['item_text'], list=list_)
+            item.full_clean()
+            item.save()
+            return redirect(f'/lists/{list_.id}/')
+        except ValidationError:
+            item.delete()
+            error = "You can't have an empty list item"
+    return render(request, 'list.html', {'list': list_, 'error': error})
